+++ packages/adapters/persistence/supabase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @multitenantkit/adapter-persistence-supabase
 
+## 0.2.4
+
+### Patch Changes
+
+- refactor: change default user table from auth.users to public.profiles
+- Updated dependencies
+  - @multitenantkit/domain@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes